cleanup
shuffle / fix xml doc data
